Move search results/mock data to a separate component
